Implement optimized versions of naive surface nets
We add two potentially faster versions of the naive surface nets algorithm. Knowing that the base naive surface nets algorithm walks the whole regular grid voxel by voxel, we can pick the longest axis of our regular grid and process slices of the grid, in other words squares, along that axis, in parallel. The second optimized version of surface nets uses a hint, which is a 3d point location that is assumed to be close to the surface. Using the hint, we perform breadth first search around the voxel containing that hint to search for an active cube. Once we find one, we then only follow along the neighboring cubes with bipolar edges and this version of surface nets thus lets us iterate over only active cubes instead of the whole 3d grid.